refactor(contacts): share page props across platforms
