+++ features/flow/contacts/src/list/native.ts
@@ -1,4 +1,4 @@
 export { ContactsPage } from "./components/ContactsPage/ContactsPage.native";
 export { ContactsAddContactHeaderButton } from "./components/ContactsPage/ContactsAddContactHeaderButton.native";
 export type { ContactsAddContactHeaderButtonProps } from "./components/ContactsPage/ContactsAddContactHeaderButton.native";
-export type { ContactsPageNativeLabels, ContactsPageNativeProps } from "./types.native";
+export type { ContactsPageNativeProps } from "./types";
